Darn, webpack doesn't play well either with using `import` for JSON instead of `require`, nor with sharing common code between server and client.
