+++ build.js
@@ -77,15 +77,15 @@ fs.writeFileSync(wayfireIniPath,lines.join("\n")+"\n")}catch(e){console.error(ch
 process_util_1.spawn)("pkill",uid,["-o",chromium]),spin,process_util_1.ErrorMode.NO_ERRORS),await(0,process_util_1.monitorProcess)((0,process_util_1.spawn)("pkill",uid,["-o",chromium.substr(0,15)]),spin,process_util_1.ErrorMode.NO_ERRORS),await(0,process_util_1.sleep)(500,spin);const args=launchChromium.split(/\s/).slice(1);args.splice(args.length-1,0,`--user-data-dir='${userHome}'`),setTimeout((()=>process.exit(0)),5e3),await(0,process_util_1.monitorProcess)((0,process_util_1.spawn)(chromium,uid,args,{
 detached:!0}))}doReboot&&(console.log(chalk.cyan(sol+"- Rebooting system in 5 seconds... -")),showStep(),write("Press any key to stop reboot:"+trailingSpace),await(0,process_util_1.sleep)(5e3,spin,!0)?console.log():(console.log(),(0,child_process_1.exec)("reboot"))),process.exit(0)}catch(err){console.log(backspace+chalk.red("✘")),console.error(err),process.exit(1)}})()},677:function(__unused_webpack_module,exports,__webpack_require__){var __importDefault=this&&this.__importDefault||function(mod){
 return mod&&mod.__esModule?mod:{default:mod}};Object.defineProperty(exports,"__esModule",{value:!0}),
-exports.safeCompareVersions=exports.setAlertId=exports.checksum53=exports.alertCleanUp=exports.filterError=exports.unref=exports.timeStamp=exports.escapeForRegex=exports.checkForecastIntegrity=exports.getRemoteAddress=exports.hasGps=exports.roughDistanceBetweenLocationsInKm=exports.splitIpAndPort=exports.normalizePort=exports.autoInHg=exports.autoHpa=exports.hpaToInHg=exports.inHgToHpa=exports.cmToInches=exports.milesToKm=exports.inchesToCm=exports.cToF=exports.fToC=exports.stdDev=exports.average=exports.jsonOrJsonp=exports.noCache=void 0
+exports.safeCompareVersions=exports.setAlertId=exports.checksum53=exports.alertCleanUp=exports.filterError=exports.unref=exports.timeStamp=exports.escapeForRegex=exports.checkForecastIntegrity=exports.getRemoteAddress=exports.hasGps=exports.roughDistanceBetweenLocationsInKm=exports.splitIpAndPort=exports.normalizePort=exports.autoInHg=exports.autoHpa=exports.hpaToInHg=exports.inHgToHpa=exports.milesToKm=exports.inchesToCm=exports.fToC=exports.stdDev=exports.average=exports.jsonOrJsonp=exports.noCache=void 0
 ;const math_1=__webpack_require__(72),process_util_1=__webpack_require__(990),util_1=__webpack_require__(714),compare_versions_1=__importDefault(__webpack_require__(323));function average(values){return values.reduce(((sum,value)=>sum+value))/values.length}function inHgToHpa(p){return 33.864*p}function hpaToInHg(p){return p/33.864}async function hasCommand(command){return!!(await(0,process_util_1.monitorProcess)((0,process_util_1.spawn)("which",[command]),null,process_util_1.ErrorMode.ANY_ERROR)).trim()}
 exports.noCache=function(res){res.header("Cache-Control","private, no-cache, no-store, must-revalidate"),res.header("Expires","-1"),res.header("Pragma","no-cache")},exports.jsonOrJsonp=function(req,res,data){req.query.callback?res.jsonp(data):res.json(data)},exports.average=average,exports.stdDev=function(values){const avg=average(values),squaredDiffs=values.map((value=>{const diff=avg-value;return diff*diff}));return Math.sqrt(average(squaredDiffs))},exports.fToC=function(f){return(f-32)/1.8},
-exports.cToF=function(c){return 1.8*c+32},exports.inchesToCm=function(i){return 2.54*i},exports.milesToKm=function(m){return 1.609344*m},exports.cmToInches=function(cm){return cm/2.54},exports.inHgToHpa=inHgToHpa,exports.hpaToInHg=hpaToInHg,exports.autoHpa=function(p){return p>100?p:inHgToHpa(p)},exports.autoInHg=function(p){return p<100?p:hpaToInHg(p)},exports.normalizePort=function(val){const port=parseInt(val,10);return isNaN(port)?val:port>=0&&port},
-exports.splitIpAndPort=function(ipWithPossiblePort,defaultPort){if(!ipWithPossiblePort)return[void 0,defaultPort];let $=/^\[(.+)]:(\d+)$/.exec(ipWithPossiblePort);return $?[$[1],Number($[2])]:($=/^([^[:]+):(\d+)$/.exec(ipWithPossiblePort),$?[$[1],Number($[2])]:[ipWithPossiblePort,defaultPort])},exports.roughDistanceBetweenLocationsInKm=function(lat1,long1,lat2,long2){let deltaRad=(0,math_1.acos)((0,math_1.sin_deg)(lat1)*(0,math_1.sin_deg)(lat2)+(0,math_1.cos_deg)(lat1)*(0,math_1.cos_deg)(lat2)*(0,
-math_1.cos_deg)(long1-long2));for(;deltaRad>math_1.PI;)deltaRad-=math_1.PI;for(;deltaRad<-math_1.PI;)deltaRad+=math_1.PI;return 6378.14*deltaRad},exports.hasGps=async function(){return await hasCommand("gpspipe")||await hasCommand("ntpq")},exports.getRemoteAddress=function(req){return req.headers["x-real-ip"]||req.socket.remoteAddress},exports.checkForecastIntegrity=function(forecast,currentOnly=!1){return(null==forecast?void 0:forecast.currently)&&(0,
-util_1.isNumber)(forecast.currently.temperature)&&(currentOnly||forecast.hourly&&forecast.hourly.length>23&&(0,util_1.isNumber)(forecast.hourly[0].temperature)&&forecast.daily&&forecast.daily.data&&forecast.daily.data.length>3&&(0,util_1.isNumber)(forecast.daily.data[0].time))};const charsNeedingRegexEscape=/[-[\]/{}()*+?.\\^$|]/g;function checksum53(s,seed=0){let h1=3735928559^seed,h2=1103547991^seed;for(let ch,i=0;i<s.length;++i)ch=s.charCodeAt(i),h1=Math.imul(h1^ch,2654435761),
-h2=Math.imul(h2^ch,1597334677);return h1=Math.imul(h1^h1>>>16,2246822507)^Math.imul(h2^h2>>>13,3266489909),h2=Math.imul(h2^h2>>>16,2246822507)^Math.imul(h1^h1>>>13,3266489909),(4294967296*(2097151&h2)+(h1>>>0)).toString(16).toUpperCase().padStart(14,"0")}exports.escapeForRegex=function(s){return s.replace(charsNeedingRegexEscape,"\\$&")},exports.timeStamp=function(){return"["+(new Date).toISOString()+"]"},exports.unref=function(timer){return(null==timer?void 0:timer.unref)&&timer.unref(),timer},
-exports.filterError=function(error){var _a;return(error=null!==(_a=null==error?void 0:error.message)&&void 0!==_a?_a:null==error?void 0:error.toString())&&error.replace(/^\s*Error:\s*/i,"")},exports.alertCleanUp=function(alertText){if(!/[\u0100-\uFFFF]/.test(alertText)&&/[\u00C0-\u00DF][\u0080-\u00BF]|([\u00E0-\u00EF][\u0080-\u00BF]{2})|([\u00F0-\u00F7][\u0080-\u00BF]{3})/.test(alertText)){const altText=Buffer.from(alertText,"latin1").toString("utf8");altText.length<alertText.length&&(alertText=altText)}
+exports.inchesToCm=function(i){return 2.54*i},exports.milesToKm=function(m){return 1.609344*m},exports.inHgToHpa=inHgToHpa,exports.hpaToInHg=hpaToInHg,exports.autoHpa=function(p){return p>100?p:inHgToHpa(p)},exports.autoInHg=function(p){return p<100?p:hpaToInHg(p)},exports.normalizePort=function(val){const port=parseInt(val,10);return isNaN(port)?val:port>=0&&port},exports.splitIpAndPort=function(ipWithPossiblePort,defaultPort){if(!ipWithPossiblePort)return[void 0,defaultPort]
+;let $=/^\[(.+)]:(\d+)$/.exec(ipWithPossiblePort);return $?[$[1],Number($[2])]:($=/^([^[:]+):(\d+)$/.exec(ipWithPossiblePort),$?[$[1],Number($[2])]:[ipWithPossiblePort,defaultPort])},exports.roughDistanceBetweenLocationsInKm=function(lat1,long1,lat2,long2){let deltaRad=(0,math_1.acos)((0,math_1.sin_deg)(lat1)*(0,math_1.sin_deg)(lat2)+(0,math_1.cos_deg)(lat1)*(0,math_1.cos_deg)(lat2)*(0,math_1.cos_deg)(long1-long2));for(;deltaRad>math_1.PI;)deltaRad-=math_1.PI;for(;deltaRad<-math_1.PI;)deltaRad+=math_1.PI
+;return 6378.14*deltaRad},exports.hasGps=async function(){return await hasCommand("gpspipe")||await hasCommand("ntpq")},exports.getRemoteAddress=function(req){return req.headers["x-real-ip"]||req.socket.remoteAddress},exports.checkForecastIntegrity=function(forecast,currentOnly=!1){return(null==forecast?void 0:forecast.currently)&&(0,util_1.isNumber)(forecast.currently.temperature)&&(currentOnly||forecast.hourly&&forecast.hourly.length>23&&(0,
+util_1.isNumber)(forecast.hourly[0].temperature)&&forecast.daily&&forecast.daily.data&&forecast.daily.data.length>3&&(0,util_1.isNumber)(forecast.daily.data[0].time))};const charsNeedingRegexEscape=/[-[\]/{}()*+?.\\^$|]/g;function checksum53(s,seed=0){let h1=3735928559^seed,h2=1103547991^seed;for(let ch,i=0;i<s.length;++i)ch=s.charCodeAt(i),h1=Math.imul(h1^ch,2654435761),h2=Math.imul(h2^ch,1597334677);return h1=Math.imul(h1^h1>>>16,2246822507)^Math.imul(h2^h2>>>13,3266489909),
+h2=Math.imul(h2^h2>>>16,2246822507)^Math.imul(h1^h1>>>13,3266489909),(4294967296*(2097151&h2)+(h1>>>0)).toString(16).toUpperCase().padStart(14,"0")}exports.escapeForRegex=function(s){return s.replace(charsNeedingRegexEscape,"\\$&")},exports.timeStamp=function(){return"["+(new Date).toISOString()+"]"},exports.unref=function(timer){return(null==timer?void 0:timer.unref)&&timer.unref(),timer},exports.filterError=function(error){var _a
+;return(error=null!==(_a=null==error?void 0:error.message)&&void 0!==_a?_a:null==error?void 0:error.toString())&&error.replace(/^\s*Error:\s*/i,"")},exports.alertCleanUp=function(alertText){if(!/[\u0100-\uFFFF]/.test(alertText)&&/[\u00C0-\u00DF][\u0080-\u00BF]|([\u00E0-\u00EF][\u0080-\u00BF]{2})|([\u00F0-\u00F7][\u0080-\u00BF]{3})/.test(alertText)){const altText=Buffer.from(alertText,"latin1").toString("utf8");altText.length<alertText.length&&(alertText=altText)}
 let alert=alertText.trim().replace(/(?<=issued an?)\n\n?\* /g," ").replace(/\bfor\.\.\.\n.*?\n\n?\* /s,(match=>{let goodLines=!0;const lines=match.split("\n").filter((line=>!!line)).slice(1,-1).map((line=>line.endsWith("...")?"• "+line.trim().slice(0,-3):(goodLines=!1,line)));return goodLines?"for:\n\n"+lines.join("\n")+"\n\n* ":match
 })).replace(/\.\.\.\n\*\s+([A-Z]{3,})\.\.\./g,".\n\n• $1: ").replace(/^((\* )?(WHAT|WHERE|WHEN|IMPACTS?|HAZARD|SOURCE))\.\.\./gm,"\n• $3: ").replace(/^- (?=\w)/gm,"  ◦ ").replace(/([12]?\d)([0-5]\d) ([AP]M)/g,"$1:$2 $3").replace(/^(.*)\.\.\.\s{3,}(.*?)(\.\.\.)?\n/gm,((match,$1,$2,$3)=>match.substring(0,35).endsWith("   ")?$3?`${$1}, ${$2}, `:`${$1}, ${$2} `:match)).replace(/^(?<!.*\.\.\.\s{3,}.*)\.\.\.\n/gm,":\n").replace(/(?<!\.)\.\n\.\.\./g,".\n\n").replace(/\.\.\.\n/g,":\n\n").replace(/^\.(?=\w)/gm,"").replace(/^[ \t]*\.{3,}/gm,"").replace(/&&\s*$/,"").replace(/&&/g,"⚠︎").replace(/\n{3,}/g,"\n\n").trim()
 ;return alert.startsWith("•")&&(alert="\n\n"+alert),alert},exports.checksum53=checksum53,exports.setAlertId=function(alert){return alert.id=checksum53(`${alert.title}\t${alert.description}\t${alert.severity}`),alert},exports.safeCompareVersions=function(firstVersion,secondVersion,operatorOrDefValue,defValue=!1){try{return(0,
