+++ raspberry_pi_setup/awc-server-error.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang=en>
+<head>
+  <title>Astronomy/Weather Clock</title>
+  <style>
+    div {
+      color: #EB3;
+      font: bold 24px Arial, sans-serif;
+      position: relative;
+      text-align: center;
+      top: 25vh;
+      width: 100%;
+    }
+  </style>
+</head>
+<body>
+<div>Astronomy/Weather Clock server failed to start</div>
+</body>
+</html>
