+++ server/src/shared-types.ts
@@ -1,4 +1,5 @@
-import { version } from '../package.json';
+// Using `require` causes a warning, but using `import` instead causes webpack to fail.
+const { version } = require('../package.json');
 
 // It's annoying that TypeScript doesn't itself provide a way to create a runtime list of
 // the allowed keys for an interface, so for now, I'll have to settle for repeating things.
