+++ src/settings-dialog.ts
@@ -9,12 +9,11 @@ import {
 import { AWC_VERSION, AwcDefaults } from '../server/src/shared-types';
 import {
   adjustCityName, ClickishEvent, decrementDialogCounter, domAlert, domConfirm, getJson, getText, incrementDialogCounter,
-  popKeydownListener, pushKeydownListener
+  popKeydownListener, pushKeydownListener, safeCompareVersions
 } from './awc-util';
 import { abs, floor, mod } from '@tubular/math';
 import { beep, clone, eventToKey, htmlEscape, isEqual, isFirefox, noop, toBoolean, toNumber } from '@tubular/util';
 import ttime, { DateTime, isValidDate_SGC } from '@tubular/time';
-import { safeCompareVersions } from '../server/src/shared-utils';
 
 const ERROR_BACKGROUND = '#FCC';
 const WARNING_BACKGROUND = '#FFC';
